pyproject.toml
[tool.poetry]
name = "az-upload-folder-to-blob"
version = "0.1.0"
description = ""
authors = ["Cristobal Villanueva <cristobaljvp@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
azure-storage-blob = "^12.14.1"
colorama = "^0.4.6"
aiohttp = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
az-blob-upload = "app.main:main"
